Move some stuff around to untangle Cauldron parts from other parts
